All modifications to generate badges manually
