--- .codeclimate.yml
@@ -1,2 +1,0 @@
-exclude_paths:
-- "test_pylint_server.py"
