--- .travis.yml
@@ -1,18 +1,0 @@
-sudo: false
-language: python
-python:
-  - "2.7"
-
-# command to install dependencies
-install:
-  - "pip install -r requirements-test.txt" 
-  - "pip install coveralls"
-  - "pip install pylint"
-
-# command to run tests
-script: nosetests --with-coverage --cover-erase --cover-inclusive --cover-package=pylint_server
-
-after_success:
-  - coveralls
-  - pylint --output-format=html pylint_server > /tmp/pylint-report.html
-  - curl -v -m 120 -X POST -F travis-job-id=$TRAVIS_JOB_ID -F pylint-report=@/tmp/pylint-report.html https://pylint.desmondrivet.com/reports
