--- requirements-deploy.txt
@@ -1,8 +1,0 @@
-Flask==0.10.1
-itsdangerous==0.24
-Jinja2==2.8
-MarkupSafe==0.23
-requests==2.7.0
-TravisPy==0.3.3
-Werkzeug==0.10.4
--e git+https://github.com/drivet/pylint-server.git#egg=pylint-server
