--- requirements-test.txt
@@ -1,5 +1,0 @@
-# a testing only requirements file for use with travis
--e .
-nose==1.3.4
-Flask-Testing==0.4.2
-coverage==3.7.1
